genai-assignment/requirements.txt

# Core Web & API
fastapi>=0.110.0
uvicorn>=0.27.0

# UI
streamlit>=1.31.0

# HTTP Client
requests>=2.31.0

# Data & Validation
pydantic>=2.5.0

# LLM / AI
openai>=1.3.0

# OCR / Document Processing 
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0

# Utilities
python-dotenv>=1.0.0
